+++ organizers/urls.py
@@ -4,7 +4,6 @@ from organizers import views
 
 urlpatterns = [
     url(r'^hacker/review/$', views.ReviewApplicationView.as_view(), name='review'),
-    url(r'^hacker/ranking/$', views.RankingView.as_view(), name='ranking'),
     url(r'^hacker/(?P<id>[\w-]+)$', views.ApplicationDetailView.as_view(), name="app_detail"),
     url(r'^hacker/all/$', views.ApplicationsListView.as_view(), name="app_list"),
     url(r'^hacker/all/invite/$', views.InviteListView.as_view(), name="invite_list"),
