+++ reimbursement/templates/reimbursement_send_table.html
@@ -10,5 +10,6 @@
         <button name="no_reimb" class="btn btn-danger btn-block"
                 value="no_reimb">Send no reimbursement
         </button>
+        <a href="{% url 'reimbursement_list' %}" class="btn-block btn btn-default">< Back</a>
     </div>
 {% endblock %}
